started working on test
