+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
+// import reactLogo from './assets/react.svg'
+// import viteLogo from '/vite.svg'
 import './App.css'
 import { PhotoViewer } from './PhotoViewer/PhotoViewer'
 import { ImageSelector } from './ImageSelector/ImageSelector'
